refactor(player-portal): decompose Character tab into stat-block components
Splits the 946-line Character.tsx into focused single-responsibility files
under components/tabs/character/, keeping zero behaviour change.
