+++ apps/player-portal/src/components/tabs/character/index.ts
@@ -1,0 +1,1 @@
+export { Character } from './Character';
